updatedBy and updatedAt update

+++ src/main/java/com/example/socapplication/model/dto/adminPermissionDto/AddAdminPermission.java
@@ -2,5 +2,5 @@ package com.example.socapplication.model.dto.adminPermissionDto;
 
 public record AddAdminPermission(
         Long userId,
-        Integer permissionId
+        Long permissionId
 ) {}
