+++ src/main/java/com/example/socapplication/repository/PermissionRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PermissionRepository extends JpaRepository<Permission, Integer> {}
+public interface PermissionRepository extends JpaRepository<Permission, Long> {}
